third_party/WebKit/LayoutTests/flag-specific/enable-blink-features=LayoutNG/external/wpt/css/css-shapes/spec-examples/shape-outside-019-expected.txt

This is a testharness.js-based test.
FAIL CSS Test: Shape from image - alpha channel, opacity > 0 + shape-margin Uncaught Error: assert_approx_equals: Line 0 is positioned properly expected 242 +/- 2 but got 234
Harness: the test ran to completion.
